Refactor CSS imports to use centralized Styles directory and add theme color meta tag
