+++ index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="icon" type="image/svg+xml" href="/icon.svg" />
+    <meta name="theme-color" content="#371443" />
     <title>Byte de Sabor</title>
     <link rel="stylesheet" type="text/css" href="/index.css" />
   </head>
