+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import './App.css';
+import './Styles/App.css';
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Home from "./pages/Home";
 import Pedidos from "./pages/Pedidos";
