+++ src/components/Header.jsx
@@ -3,7 +3,7 @@ import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import 'bootstrap-icons/font/bootstrap-icons.css';
 import Logo from '../Assets/Img/Logo.svg'; // Corrija o caminho se necessário
-import './Header.css';
+import '../Styles/Header.css';
 
 const Header = () => {
     const [isOpen, setIsOpen] = useState(false);
